Use better words
