+++ client/src/shared/constants/index.ts
@@ -1,1 +1,1 @@
-export * from './FundingSpaceTypes';
+export * from './FundingSourceTimes';
