+++ client/src/shared/models/FundingSourceTime.ts
@@ -1,0 +1,8 @@
+import { FundingSource } from ".";
+import { FundingTimeInput } from "./FundingTime";
+
+export interface FundingSourceTime {
+  displayName: string;
+  fundingSources: FundingSource[];    //  All FundingSources that can have the associated FundingTimes
+  fundingTimes: FundingTimeInput[];   //  All valid FundingTimes for the associated FundingSources 
+}
