+++ client/src/shared/models/FundingTime.ts
@@ -6,3 +6,9 @@ export enum FundingTime {
   ExtendedYear = 'ExtendedYear',
   School = 'School',
 }
+
+export interface FundingTimeInput {
+  displayName: string;  //  A user-friendly version of this funding time
+  formats: string[];    //  All accepted input formats for this funding time
+  value: FundingTime;   //  The actual underlying funding time
+}
